created seedOrder.js and created api endpoint
